Simplify participant transfers
It used to be that when a participant transferred from one class
to another, all their reports data had to be moved to the new
class. No more! Now, reports.php and m_reports.php look for data
for all classes that start in the same calendar month as the
participant's current class. So when a participant transfers, no
data has to be moved. If the transfer is within the same calendar
month, the participant still sees their old data. If it's to a
different calendar month, they don't.

This change also removes old participant transfer functions from
config.php.